+++ crates/nostr/src/event/builder.rs
@@ -741,6 +741,42 @@ impl EventBuilder {
         Ok(builder)
     }
 
+    /// Voice Message
+    ///
+    /// Note: This will not add `imeta` tag ([NIP-92])
+    ///
+    /// <https://github.com/nostr-protocol/nips/blob/master/A0.md>
+    ///
+    /// [NIP-92]: https://github.com/nostr-protocol/nips/blob/master/92.md
+    #[inline]
+    pub fn voice_message<T>(voice_url: T) -> Self
+    where
+        T: Into<Url>,
+    {
+        EventBuilder::new(Kind::VoiceMessage, voice_url.into().as_str())
+    }
+
+    /// Voice Message Reply
+    ///
+    /// Note: This will not add `imeta` tag ([NIP-92])
+    ///
+    /// <https://github.com/nostr-protocol/nips/blob/master/A0.md>
+    ///
+    /// [NIP-92]: https://github.com/nostr-protocol/nips/blob/master/92.md
+    #[inline]
+    pub fn voice_message_reply<T>(
+        voice_url: T,
+        root: Option<CommentTarget<'_>>,
+        parent: CommentTarget<'_>,
+    ) -> Self
+    where
+        T: Into<Url>,
+    {
+        EventBuilder::new(Kind::VoiceMessageReply, voice_url.into().as_str())
+            .tags(root.map(|c| c.as_vec(true)).unwrap_or_default())
+            .tags(parent.as_vec(false))
+    }
+
     /// Add reaction (like/upvote, dislike/downvote or emoji) to an event
     ///
     /// <https://github.com/nostr-protocol/nips/blob/master/25.md>