+++ crates/nostr/CHANGELOG.md
@@ -35,6 +35,7 @@
 - Add repository state announcements kind `Kind::RepoState` (30618) (https://github.com/rust-nostr/nostr/pull/1041)
 - Add `HEAD` tag kind (https://github.com/rust-nostr/nostr/pull/1043)
 - Add `CommentTarget::as_vec` to convert the comment target into a vector of tags (https://github.com/rust-nostr/nostr/pull/1038)
+- Support NIP-A0 (Voice Messages) (https://github.com/rust-nostr/nostr/pull/1032)
 
 ### Changed